pyproject.toml
[tool.poetry]
name = "ecg-classifier"
version = "0.1.0"
description = "training pipeline for ECG image classification"
authors = ["grshaulov@gmail.com"]
readme = "README.md"
packages = [{ include = "ecg_classifier" }]

[[tool.poetry.source]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
priority = "explicit"

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^1.26.0"
pillow = "^10.0.0"

pytorch-lightning = "^2.4.0"
torchmetrics = "^1.4.0"
timm = "^1.0.0"

hydra-core = "^1.3.2"
fire = "^0.6.0"

mlflow = "^2.14.0"

requests = "^2.32.3"

dvc = { version = "^3.50.0", extras = ["s3"] }

fastapi = "0.115.0"
uvicorn = { version = "0.30.6", extras = ["standard"] }
celery = "5.4.0"
redis = "5.0.8"
pyjwt = "2.9.0"
python-multipart = "0.0.9"
pydantic-settings = "2.5.2"
sqlalchemy = "2.0.36"
psycopg = { version = "3.2.3", extras = ["binary"] }
alembic = "1.13.3"
[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.0"
ruff = "^0.6.8"
pytest = "^8.2.0"

[tool.poetry.group.cpu.dependencies]
torch = "2.5.1"
torchvision = "0.20.1"
torchaudio = "2.5.1"

[tool.poetry.group.gpu]
optional = true

[tool.poetry.group.gpu.dependencies]
torch = { version = "2.5.1+cu121", source = "pytorch-cu121" }
torchvision = { version = "0.20.1+cu121", source = "pytorch-cu121" }
torchaudio = { version = "2.5.1+cu121", source = "pytorch-cu121" }

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"